pyproject.toml
[project]
name = "pyquerymatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "thewisenerd", email = "thewisenerd@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyyaml>=6.0.2",
]
